更新プレイヤーと敵

+++ つっぱり超相撲ver0.12/game.cpp
@@ -297,9 +297,9 @@ void CGame::Update(void)
 			//‰–
 			moveRand.x = sinf((rand() % 628) / 100.0f) * ((rand() % 3 + 1));
 			moveRand.y = cosf((rand() % 628) / 20.0f) * ((rand() % 6 + 3));
-			moveRand.z = ((rand() % 7 + 3));
+			moveRand.z = (float)((rand() % 7 + 3));
 			//moveRand.x = rand() % 2 - 2;
-			PosRand.x = rand() % 300 - 300;
+			PosRand.x = (float)(rand() % 300 - 300);
 
 			CEffect3D::Create(D3DXVECTOR3(0.0f, 100.0f, 0.0f), D3DXVECTOR3(moveRand.x, moveRand.y, -moveRand.z), D3DXCOLOR(1, 1, 1, 1),
 				6, 6, 1, 200, CLoad::TEXTURE_EFFECT_NORMAL000);
